+++ src/main/java/me/kall/itisme/ItIsMe.java
@@ -1,5 +1,7 @@
 package me.kall.itisme;
 
+import me.kall.duplicationless.util.Mods;
+import me.kall.itisme.integration.SodiumIntegration;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
@@ -12,5 +14,7 @@ public final class ItIsMe {
     public ItIsMe(FMLJavaModLoadingContext context) {
         if (!FMLLoader.getDist().isClient()) return;
         context.registerConfig(ModConfig.Type.CLIENT, ItIsMeConfig.LIGHT_TWEAK_CONFIG);
+
+        if (Mods.isLoaded("embeddium")) SodiumIntegration.register();
     }
 }