+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 # 1.0.1
 - Fix shadow outline rendering of baby entities on 1.21.1
 - Fix shadow outline is rendered on the grass level.
+# 1.1.0
+- Embeddium video setting options integration
+- Config option to toggle entity outline and entity shadow outline